bgituQuizzTG/requirements.txt
aiogram
requests
python-dotenv 